Normative: Always report assignment failure to canonical numeric string properties of immutable-backed TypedArray instances
Note that this includes non-integer-index but still canonical numeric
properties such as "-0", "-1", "1.2", and String(2**53) =
"9007199254740992", but does not include non-canonical-numeric
properties such as "1.20", "1e3", and "0x20".

Ideally, TypedArrays would restrict special treatment of canonical
numeric string properties to the strict subset that are integer indices
(aligning with similar treatment in String), but that is out of scope
for this commit.