Normative: Report assignment failure to canonical numeric properties of immutable-backed TypedArrays
Note that this includes non-integer-index but still canonical numeric
properties such as "-0", "-1", "1.2", and String(2**53) =
"9007199254740992", but does not include non-canonical-numeric
properties such as "1.20", "1e3", and "0x20", which are not subject to
special treatment and therefore can be set as on any ordinary object.

Ideally, TypedArrays would restrict special treatment of canonical
numeric string properties to the strict subset that are integer indices
(round-trippable integers in the inclusive interval from 0 to 2**53 - 1),
but that is out of scope for this commit.